Fix logo centering and content
- Remove `.gridCell--row` node around `.pageHeader`. What we want to do
with `.pageHeader` is now a little more specific, so relying on the
current grid system here complicates things a bit—easier to ditch it
for this special case.
- Add `.sanskrit` node to `beginners-yoga-course.html`.
- Update the CSS to always center `.pageHeader`. The most important
stuff here are the flexbox styles which are hopefully pretty clear. The
`width` was necessary on `.sanskrit` to prevent it from disappearing.
It worked before because it was a `div` which is by default `width:
100%` but stopped working once I switched `.pageHeader` to flexbox
because it no longer had that default width since it's a flex child.